pyproject.toml
[project]
name = "mcp-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=1.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "streamlit>=1.46.1",
    "streamlit-authenticator>=0.4.2",
    "uvicorn>=0.35.0",
]
